+++ lib/libyomu/collection.ml
@@ -72,18 +72,99 @@ module Normal = struct
         files
     in
     ()
+
+  (**
+      [named_archive_of_name root comic_name file_name]  
+    *)
+  let named_archive_of_name root comic_name file_name =
+    let ( // ) = Filename.concat in
+    let archive_path = root // comic_name // file_name in
+    let name = Printf.sprintf "%s-%s" comic_name file_name in
+    Item.{ archive_path; name }
+
+  let yomu_comics () = Array.to_list @@ Sys.readdir App.yomu_comics
+
+  (**
+        [read_comic_dir name] read the content of [Libyomu.App.yomu_comics/name]
+    *)
+  let read_comic_dir name =
+    let full_path = Filename.concat App.yomu_comics name in
+    List.filter (Fun.negate @@ String.starts_with ~prefix:".")
+    @@ Array.to_list @@ Sys.readdir full_path
+
+  (**
+      [matchesp regex s] matches the the comics with the name [name] and converts all comics to 
+      [Item.named_archive]. If [regex], [name] is treated as a regular expression
+  *)
+  let matchesp regex s =
+    let regexp =
+      match regex with true -> Str.regexp s | false -> Str.regexp_string s
+    in
+    yomu_comics ()
+    |> List.filter_map (fun name ->
+           match Str.string_match regexp name 0 with
+           | true when not @@ String.starts_with ~prefix:"." name ->
+               let elts = read_comic_dir name in
+               Option.some
+               @@ List.sort NamedArchive.compare_named_archive
+               @@ List.map (named_archive_of_name App.yomu_comics name) elts
+           | false | true ->
+               None
+       )
+    |> List.flatten
+
+  (**
+      [matchesip regex index name] matches the the comics with the name [name] and its index [index] and converts all comics to 
+      [Item.named_archive]. If [regex], [name] is treated as a regular expression
+  *)
+  let matchesip regex index name =
+    let regexp =
+      match regex with
+      | true ->
+          Str.regexp name
+      | false ->
+          Str.regexp_string name
+    in
+    let ( // ) = Filename.concat in
+    yomu_comics ()
+    |> List.filter_map (fun name ->
+           match Str.string_match regexp name 0 with
+           | true when not @@ String.starts_with ~prefix:"." name ->
+               Some (name, read_comic_dir name)
+           | false | true ->
+               None
+       )
+    |> List.map (fun (comic_name, elts) ->
+           let archive_path =
+             List.filter_map
+               (fun file ->
+                 match int_of_string_opt @@ Filename.remove_extension file with
+                 | Some n when n = index ->
+                     let path = App.yomu_comics // comic_name in
+                     let name = Printf.sprintf "%s-%s" comic_name file in
+                     let archive_path = path // file in
+                     let ar = Item.{ archive_path; name } in
+                     Option.some ar
+                 | None | Some _ ->
+                     None
+               )
+               elts
+           in
+           archive_path
+       )
+    |> List.flatten
 end
 
 module Encrypted = struct
   let check_serie_exist comic_name syomurc () =
-    match Comic.Syomu.serie_exists comic_name syomurc with
+    match Syomu.serie_exists comic_name syomurc with
     | true ->
         ()
     | false ->
         failwith "S: Serie doesnt exist"
 
   let check_duplicate comic_name volume syomurc () =
-    match Comic.Syomu.exists volume comic_name syomurc with
+    match Syomu.exists volume comic_name syomurc with
     | false ->
         ()
     | true ->
@@ -115,17 +196,17 @@ module Encrypted = struct
     let () = check_duplicate comic_name index syomurc () in
     let name, extension = dname comic_name index comic_archive in
     let digest_name = Util.Hash.hash_name ~name ~extension in
-    let item = Comic.Syomu.create_item digest_name comic_name index in
+    let item = Syomu.create_item digest_name comic_name index in
     let content = Util.Io.content_filename comic_archive () in
     let content_outfile = App.yomu_hidden_comics // digest_name in
     let _encrypted_content =
       Encryption.encrypt ~where:content_outfile ~key ~iv:item.iv content ()
     in
-    let syomurc = Comic.Syomu.add item syomurc in
+    let syomurc = Syomu.add item syomurc in
     syomurc
 
   let add_multiples ~key ~existing ~comic_name indexed_archives =
-    let syomurc = Comic.Syomu.decrypt ~key () in
+    let syomurc = Syomu.decrypt ~key () in
     let syomurc =
       indexed_archives
       |> List.fold_left
@@ -134,7 +215,7 @@ module Encrypted = struct
            )
            syomurc
     in
-    let _ = Comic.Syomu.encrypt ~key syomurc () in
+    let _ = Syomu.encrypt ~key syomurc () in
     ()
 end